Bug #531643, update jpeg codec to support encoding 21M
[root cause  ] the calculating of total_mcu_num is not correct.
[changes     ] jpeg codec
[side effects] null
[self test   ]<sc9860, taking picture with resolution of 21M works>:
[reviewers   ] Xiaowei Luo
[change_type ] bringup --> sprd_chip_support
[tag_product ] specific

+++ jpeg/inc/jpegenc_frame.h
@@ -27,7 +27,7 @@
 #endif
 
 		
-PUBLIC void START_HW_ENCODE(uint16 total_mcu_num);
+PUBLIC void START_HW_ENCODE(uint32 total_mcu_num);
 
 
 /**---------------------------------------------------------------------------*
